+++ src/audio/FFTProcessor.cpp
@@ -1,3 +1,12 @@
+/**
+ * @file FFTProcessor.cpp
+ * @brief KissFFT-backed float FFT implementation.
+ *
+ * Thin wrapper around kiss_fft that manages config allocation/deallocation via
+ * RAII (KissCfg) and exposes forward transform, normalised inverse, and
+ * magnitude helpers — all operating on float precision.
+ */
+
 #include "audio/FFTProcessor.hpp"
 #include "core/Errors.hpp"
 
@@ -20,7 +29,7 @@ struct KissCfg {
 
 namespace fft {
 
-std::vector<std::complex<double>> transform(const std::vector<float> &input) {
+std::vector<std::complex<float>> transform(const std::vector<float> &input) {
     if (input.empty())
         throw dissonance::DspError("FFT input cannot be empty");
 
@@ -35,13 +44,13 @@ std::vector<std::complex<double>> transform(const std::vector<float> &input) {
 
     kiss_fft(cfg.cfg, in.data(), out.data());
 
-    std::vector<std::complex<double>> result(N);
+    std::vector<std::complex<float>> result(N);
     for (int i = 0; i < N; ++i)
-        result[i] = {static_cast<double>(out[i].r), static_cast<double>(out[i].i)};
+        result[i] = {out[i].r, out[i].i};
     return result;
 }
 
-std::vector<double> inverse(const std::vector<std::complex<double>> &spectrum) {
+std::vector<float> inverse(const std::vector<std::complex<float>> &spectrum) {
     if (spectrum.empty())
         throw dissonance::DspError("iFFT input cannot be empty");
 
@@ -50,21 +59,21 @@ std::vector<double> inverse(const std::vector<std::complex<double>> &spectrum) {
 
     std::vector<kiss_fft_cpx> in(N), out(N);
     for (int i = 0; i < N; ++i) {
-        in[i].r = static_cast<float>(spectrum[i].real());
-        in[i].i = static_cast<float>(spectrum[i].imag());
+        in[i].r = spectrum[i].real();
+        in[i].i = spectrum[i].imag();
     }
 
     kiss_fft(cfg.cfg, in.data(), out.data());
 
-    const double scale = 1.0 / static_cast<double>(N);
-    std::vector<double> result(N);
+    const float scale = 1.0f / static_cast<float>(N);
+    std::vector<float> result(N);
     for (int i = 0; i < N; ++i)
-        result[i] = static_cast<double>(out[i].r) * scale;
+        result[i] = out[i].r * scale;
     return result;
 }
 
-std::vector<double> magnitude(const std::vector<std::complex<double>> &spectrum) {
-    std::vector<double> mags(spectrum.size());
+std::vector<float> magnitude(const std::vector<std::complex<float>> &spectrum) {
+    std::vector<float> mags(spectrum.size());
     for (size_t i = 0; i < spectrum.size(); ++i)
         mags[i] = std::abs(spectrum[i]);
     return mags;
