+++ src/audio/WindowFunctions.cpp
@@ -1,7 +1,11 @@
+/**
+ * @file WindowFunctions.cpp
+ * @brief Hann and Hamming window generation and application.
+ */
+
 #include "audio/WindowFunctions.hpp"
 #include "core/Errors.hpp"
 
-#include <algorithm>
 #include <cmath>
 
 #ifndef M_PI
@@ -12,34 +16,33 @@ namespace window {
 
 namespace {
 
-std::vector<double> generateHann(size_t size) {
-    std::vector<double> w(size);
+std::vector<float> generateHann(size_t size) {
+    std::vector<float> w(size);
     if (size == 1) {
-        w[0] = 1.0;
+        w[0] = 1.0f;
         return w;
     }
     for (size_t n = 0; n < size; ++n)
-        w[n] =
-            0.5 *
-            (1.0 - std::cos(2.0 * M_PI * static_cast<double>(n) / static_cast<double>(size - 1)));
+        w[n] = 0.5f * (1.0f - std::cos(2.0f * static_cast<float>(M_PI) *
+                                        static_cast<float>(n) / static_cast<float>(size - 1)));
     return w;
 }
 
-std::vector<double> generateHamming(size_t size) {
-    std::vector<double> w(size);
+std::vector<float> generateHamming(size_t size) {
+    std::vector<float> w(size);
     if (size == 1) {
-        w[0] = 1.0;
+        w[0] = 1.0f;
         return w;
     }
     for (size_t n = 0; n < size; ++n)
-        w[n] = 0.54 -
-               0.46 * std::cos(2.0 * M_PI * static_cast<double>(n) / static_cast<double>(size - 1));
+        w[n] = 0.54f - 0.46f * std::cos(2.0f * static_cast<float>(M_PI) *
+                                         static_cast<float>(n) / static_cast<float>(size - 1));
     return w;
 }
 
 } // namespace
 
-std::vector<double> generate(Type type, size_t size) {
+std::vector<float> generate(Type type, size_t size) {
     if (size == 0)
         throw dissonance::DspError("Window size must be greater than 0");
     switch (type) {
@@ -52,11 +55,11 @@ std::vector<double> generate(Type type, size_t size) {
     }
 }
 
-void apply(std::vector<float> &samples, const std::vector<double> &coefficients) {
+void apply(std::vector<float> &samples, const std::vector<float> &coefficients) {
     if (samples.size() != coefficients.size())
         throw dissonance::DspError("Sample buffer size must match window size");
     for (size_t i = 0; i < samples.size(); ++i)
-        samples[i] *= static_cast<float>(coefficients[i]);
+        samples[i] *= coefficients[i];
 }
 
 } // namespace window
